Сделал абстрактный класс предок BasicSettings и вынес туда BeforeEach и AfterEach из всех тестовых классов
